Extract relay structs
